restructure folders and add name checker and scraper
