--- test.js
@@ -1,44 +1,0 @@
-var str = `<!DOCTYPE html>
-<html>
-  <head>
-    <meta charset="utf-8">
-      <title></title>
-    </head>
-    <body>
-      <div class="paper-stories">
-        <div class="-spr1 -shaded -center">
-          <p class="text-h3"> LIVERPOOL ECHO </p>
-          <!--<img src="/images/football/papertalk-logos-liverpoolecho.gif" alt="liverpool echo">-->
-        </div>
-          <ul class="list -bullet">
-            <li class="-spr1 item">West Brom manager Tony Pulis says he let Brendan Galloway return to Everton because he could not promise the defender regular football at the Hawthorns.</li>
-            <li class="-spr1 item">Lionel Messi is against Barcelona re-signing Everton forward Gerard Deulofeu this summer, say reports in Spain.</li>
-            <li class="-spr1 item">Liverpool are no longer in the running to land unsettled Madrid playmaker James Rodriguez next season, according to reports in Spain.</li>
-            <li class="-spr1 item">Lucas Leiva has still to decide if he will stay at Liverpool next season, with the Brazilian understood to want more regular first-team football at Anfield if he does opt to stay.</li>
-          </ul>
-      </div>
-        <div class="paper-stories">
-          <div class="-spr1 -shaded -center">
-            <p class="text-h3"> LONDON EVENING STANDARD </p>
-            <!--<img src="/images/football/papertalk-logos-londoneveningstandard.gif" alt="london evening standard">-->
-            </div>
-            <ul class="list -bullet">
-              <li class="-spr1 item">Departing Chelsea captain John Terry insists he still has plenty to offer and will play on at a new club next season.</li>
-              <li class="-spr1 item">West Brom are John Terry's most likely next destination if he remains in the Premier League, according to the latest betting odds.</li>
-            </ul>
-          </div>
-          <div class="paper-stories">
-            <div class="-spr1 -shaded -center">
-              <p class="text-h3"> NEWCASTLE CHRONICLE </p>
-              <!--<img src="/images/football/papertalk-logos-newcastlechronicle.gif" alt="newcastle chronicle">-->
-              </div>
-              <ul class="list -bullet">
-                <li class="-spr1 item">Napoli are interested in signing Sunderland winger Jeremain Lens when the Dutchman's loan spell at Fenerbahce finishes at the end of the season, according to reports in the Italian media.</li>
-                <li class="-spr1 item">Sunderland boss David Moyes is believed to have been impressed with Southampton striker Sam Gallagher's displays while on loan at Blackburn this season.</li>
-                <li class="-spr1 item">And contrary to some reports this morning, the Scot's job at the Stadium of Light is not thought to be under threat, despite the Black Cats' precarious position at the foot of the Premier League.</li>
-              </ul>
-            </div>
-
-          </body>
-        </html>`
-module.exports = str;
